researches/MAI-UI/webui/requirements.txt
# MAI-UI WebUI Dependencies
# Core
Jinja2>=3.1.0
numpy>=2.0.0
openai>=1.0.0
Pillow>=10.0.0

# Web UI
gradio>=4.0.0
pyyaml>=6.0.0
requests>=2.28.0

# Data Processing
tqdm>=4.65.0

# PDF Export (optional)
reportlab>=4.0.0
